backend/requirements.txt
langgraph==0.2.*
pydantic==2.*
pyyaml==6.*
pandas==2.*
